server: Replace CryptoRandom with CryptoRngCore + SeedableRng
This patch adds a new dependency for interfacing with random number
generators, which is rand_core. rand_core contains core traits and error
types for working with random number generators. It is considered the
most mature API for random number generators and I think it has been
battle tested and will be better than anything we come up with. (unless
we find a usecase where we have to)

Signed-off-by: Ahmed Abdelraoof <ahmed.abdelraoof@huawei.com>